requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
pdfplumber
scikit-learn
python-multipart
requests